backend/app/requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
SQLAlchemy
python-dotenv
websockets
